ASoC: rt5651: Fix bias_level confusion / overcurrent detection deps
Overcurrent detection (ovcd) requires the following to be on:
1) The LDO supply
2) The micbias1 supply
3) General analog voltages such as vref aka a bias_level of standby

Before this commit deps 2. and 3. were not met (unless a stream recording
from the mic was active).

3. Is not met because rt5651_set_bias_level() was only enabling this when
reaching a bias level of prepared instead of doing this in the normal
standby bias level, which the dapm core will select as soon as any pins /
supplies are on. This commit fixes by making rt5651_set_bias_level() behave
as a normal set_bias function for other codecs and already enabling these
things at standby level.

This change to rt5651_set_bias_level() causes a problem because when jack-
detect is used the bias-level was always set to standby because of the
"JD Power" supply being force-enabled.

As the set_bias_level code already leaves the RT5651_PWR_JD_M bit on when
entering standby (now off) when jd is in use, we can simply drop the
"JD Power" supply so that the bias-level properly becomes off when nothing
is happening. For the same reason we should also not enable the LDO supply
until we actually want to do ovcd detection.

2. is fixed by simply force-enabling "micbias1" when doing ovcd.

Also make sure to turn both the micbias1 and the LDO supplies off again
when we're done, note they will only really get turned off if the ovcd
was the only user.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>