ASoC: rt5651: Improve headphone vs headset detection
Headphone vs headset detection relies on OVCD and testing on the rt5640 and
rt5651 (which seem to have the same OVCD detection) has shown that getting
OVCD detection to work reliabe on these codecs is somewhat finicky.

This commit ports my work to make this reliable on the rt5640 over to the
rt5651, making the following OVCD changes, each of which has been tested
individually and each of which has shown to be necessary on the rt5651:

1) When the mic-gnd contacts are short-circuited by a headphones plug, the
hardware periodically retries if it can apply the bias-current leading to
the OVCD status flip-flopping 1-0-1 with it being 0 about 10% of the time.
This commit enables the sticky bit for the OVCD status to deal with this.

2) When using RCCLK instead of MCLK / PLL1 the OVCD status often gets stuck
at its last value, force-enable the platform clock to avoid this.

3) The OVCD reads 0 / mic-gnd open sometimes even when short-circuited,
setting the OVCD detection threshold scale-factor to 0.75 fixes this?

This commit also changes the OVCD detection threshold to 1500uA (0.75 *
2000) because lower values lead to false-positive headphones detection,
even 1500uA leads to false-positive headphone detection when using old
Nokia phone headsets, but higher values make almost anything detected as a
headsets, so 1500uA it is.

Finally this commit changes the detection algorithm to require 5 identical
OVCD values in a row, as OVCD may bounce a bit after jack insertion.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Note this is based on the jack-detection work I've been doing for the rt5640,
I took a detour and ended up fixing this for rt5651 first, the rt5640 work
is almost ready too and I hope to post a version of that upstream soon.