power: supply: axp288_charger: Use model specific current limit for power-barrel
Some devices with an AXP288 PMIC use a power-barrel connector instead of
a micro-usb connector for charging (some have both but will only charge
through the power-barrel connector).

In this case the charger detection will always recognize the charge-port
as a Dedicated Charge Port (DCP), leading to an input-current-limit of 2A.

But the custom. device-specific adapter shipped with these devices
can usually deliver significantly more then 2A, typically 3 or 4A.

This commit adds a dmi based table to recognize such devices and to
automatically set the input-current-limit to match the adapter these
devices ship with.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>