Input: silead - support touchscreens with non 0 minimum coordinates
Some Silead bases touchscreens (depending on the firmware and/or the
regulator) report coordinates which never reach 0 along one or both
of their axis.

This commit adds support for setting a "silead,min-x" and "min-y" property
which indicates the minimum x/y values reported.

When enabled this fixes e.g. not being able to click things in the GNOME3
top-bar on same tablets.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>